+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Discriminated unions can now have nested models within them
 - Discriminated unions can now be part of nested models
 
+### Changed
+- Checkbox and switch now have a default of False rather than None
+
 ### Added
 - New regression tests